--- cookbook/rl/async_multi_lora_dapo_hparam_sweep.yaml
@@ -1,203 +1,0 @@
-runtime:
-  run_id: async_multi_lora_dapo_hparam_sweep
-  model_id: ${oc.env:MODEL_ID,ms://Qwen/Qwen3-4B}
-  mode: ray
-  # Three-GPU layout: two trainer GPUs in one SP group plus one sampler GPU.
-  model_gpus: 2
-  sampler_gpus: 1
-  sampler_tp: 1
-  sampler_max_loras: 4
-  max_staleness: 2
-  max_steps: null
-  allow_partial_rollout: false
-  rollout_max_retries: 2
-  rollout_retry_delay_s: 0.5
-  keep_adapter_versions: 2
-  output_dir: output/async_multi_lora_dapo_hparam_sweep
-
-metrics:
-  enabled: true
-  drain_interval_s: 1.0
-  queue_capacity: 10000
-  close_timeout_s: 10
-  jsonl:
-    enabled: true
-    path: outputs/async_rl/dapo_hparam_sweep_metrics.jsonl
-    summary_path: outputs/async_rl/dapo_hparam_sweep_summary.json
-    batch_size: 64
-    flush_interval_s: 2.0
-  swanlab:
-    enabled: false
-    mode: local
-    project: twinkle-rl
-    name: ${runtime.run_id}
-    log_dir: outputs/swanlab
-
-template:
-  cls: Template
-  enable_thinking: false
-
-model:
-  strategy: native_fsdp
-  fsdp_config:
-    reshard_after_forward: true
-  sequence_parallel_size: 2
-  padding_free: true
-  attn_implementation: flash_attention_2
-  max_length: 12288
-
-sampler:
-  max_model_len: 12288
-  gpu_memory_utilization: 0.8
-  max_num_seqs: 32
-  max_num_batched_tokens: 16384
-  enforce_eager: true
-
-tq:
-  polling_mode: true
-  storage_units: 2
-
-scheduler:
-  rollout: {policy: round_robin, max_consecutive_units: 1}
-  advantage: {policy: oldest_partition, max_consecutive_units: 1}
-  # Interleave hyperparameter candidates so one adapter cannot monopolize training.
-  train: {policy: round_robin, max_consecutive_units: 1}
-
-lora:
-  target_modules: all-linear
-  r: 16
-  alpha: 32
-  dropout: 0.05
-  # Used only when a context omits train.learning_rate.
-  learning_rate: 1.0e-6
-
-loss:
-  cls: GRPOLoss
-  epsilon: 0.2
-
-lora_contexts:
-  # A/B/C isolate learning rate while keeping mini_batch_size fixed at 32 samples.
-  - tenant_id: tenant_lr1e6_g4
-    training_run_id: dapo_hparam_sweep
-    adapter_name: dapo_lr1e6_g4_lora
-    reward:
-      class_path: twinkle.reward.DAPOMathReward
-      kwargs:
-        max_response_length: ${...rollout.max_tokens}
-        overlong_buffer_length: 4096
-        overlong_penalty_factor: 1.0
-        score_tail_chars: 300
-    dataset:
-      dataset_id: ${oc.env:DAPO_HPARAM_DATASET_ID,ms://BytedTsinghua-SIA/DAPO-Math-17k}
-      subset_name: default
-      split: train
-      data_num: 500
-      max_length: 4096
-      processor: DAPOMathProcessor
-    rollout:
-      batch_size: 16
-      num_generations: 8
-      max_tokens: 8192
-      temperature: 1.0
-      top_p: 0.95
-    train:
-      learning_rate: 1.0e-6
-      mini_batch_size: 32
-      micro_batch_size: 1
-      dynamic_batching: false
-
-  - tenant_id: tenant_lr5e6_g4
-    training_run_id: dapo_hparam_sweep
-    adapter_name: dapo_lr5e6_g4_lora
-    reward:
-      class_path: twinkle.reward.DAPOMathReward
-      kwargs:
-        max_response_length: ${...rollout.max_tokens}
-        overlong_buffer_length: 4096
-        overlong_penalty_factor: 1.0
-        score_tail_chars: 300
-    dataset:
-      dataset_id: ${oc.env:DAPO_HPARAM_DATASET_ID,ms://BytedTsinghua-SIA/DAPO-Math-17k}
-      subset_name: default
-      split: train
-      data_num: 500
-      max_length: 4096
-      processor: DAPOMathProcessor
-    rollout:
-      batch_size: 16
-      num_generations: 8
-      max_tokens: 8192
-      temperature: 1.0
-      top_p: 0.95
-    train:
-      learning_rate: 5.0e-6
-      mini_batch_size: 32
-      micro_batch_size: 1
-      dynamic_batching: false
-
-  - tenant_id: tenant_lr1e5_g4
-    training_run_id: dapo_hparam_sweep
-    adapter_name: dapo_lr1e5_g4_lora
-    reward:
-      class_path: twinkle.reward.DAPOMathReward
-      kwargs:
-        max_response_length: ${...rollout.max_tokens}
-        overlong_buffer_length: 4096
-        overlong_penalty_factor: 1.0
-        score_tail_chars: 300
-    dataset:
-      dataset_id: ${oc.env:DAPO_HPARAM_DATASET_ID,ms://BytedTsinghua-SIA/DAPO-Math-17k}
-      subset_name: default
-      split: train
-      data_num: 500
-      max_length: 4096
-      processor: DAPOMathProcessor
-    rollout:
-      batch_size: 16
-      num_generations: 8
-      max_tokens: 8192
-      temperature: 1.0
-      top_p: 0.95
-    train:
-      learning_rate: 1.0e-5
-      mini_batch_size: 32
-      micro_batch_size: 1
-      dynamic_batching: false
-
-  # B/D isolate train batch size while keeping learning_rate fixed at 5e-6.
-  - tenant_id: tenant_lr5e6_g16
-    training_run_id: dapo_hparam_sweep
-    adapter_name: dapo_lr5e6_g16_lora
-    reward:
-      class_path: twinkle.reward.DAPOMathReward
-      kwargs:
-        max_response_length: ${...rollout.max_tokens}
-        overlong_buffer_length: 4096
-        overlong_penalty_factor: 1.0
-        score_tail_chars: 300
-    dataset:
-      dataset_id: ${oc.env:DAPO_HPARAM_DATASET_ID,ms://BytedTsinghua-SIA/DAPO-Math-17k}
-      subset_name: default
-      split: train
-      data_num: 500
-      max_length: 4096
-      processor: DAPOMathProcessor
-    rollout:
-      batch_size: 16
-      num_generations: 8
-      max_tokens: 8192
-      temperature: 1.0
-      top_p: 0.95
-    train:
-      learning_rate: 5.0e-6
-      mini_batch_size: 128
-      micro_batch_size: 1
-      dynamic_batching: false
-
-
-# CUDA_VISIBLE_DEVICES=0,1,2 \
-# MODEL_ID=/nas/disk1/Qwen3-4B \
-# DAPO_HPARAM_DATASET_ID=/path/to/dapo_math_500.parquet \
-# TWINKLE_SEED=42 \
-# python3 cookbook/rl/async_multi_lora_grpo.py \
-#   --config cookbook/rl/async_multi_lora_dapo_hparam_sweep.yaml
