--- cookbook/rl/async_single_lora_dapo_grpo.yaml
@@ -1,133 +1,0 @@
-runtime:
-  run_id: async_single_lora_dapo_grpo
-  model_id: ${oc.env:MODEL_ID,ms://Qwen/Qwen3-4B}
-  mode: ray
-  # Same three-GPU layout as the four-context sweep: two trainer GPUs plus one sampler GPU.
-  model_gpus: 2
-  sampler_gpus: 1
-  sampler_tp: 1
-  sampler_max_loras: 4
-  max_staleness: 0
-  max_steps: null
-  allow_partial_rollout: false
-  rollout_max_retries: 2
-  rollout_retry_delay_s: 0.5
-  keep_adapter_versions: 2
-  output_dir: output/async_single_lora_dapo_grpo
-
-metrics:
-  enabled: true
-  drain_interval_s: 1.0
-  queue_capacity: 10000
-  close_timeout_s: 10
-  jsonl:
-    enabled: true
-    path: outputs/async_rl/single_lora_dapo_metrics.jsonl
-    summary_path: outputs/async_rl/single_lora_dapo_summary.json
-    batch_size: 64
-    flush_interval_s: 2.0
-  swanlab:
-    enabled: false
-    mode: local
-    project: twinkle-rl
-    name: ${runtime.run_id}
-    log_dir: outputs/swanlab
-
-template:
-  cls: Template
-  enable_thinking: false
-
-model:
-  strategy: native_fsdp
-  fsdp_config:
-    reshard_after_forward: true
-  sequence_parallel_size: 2
-  padding_free: true
-  attn_implementation: flash_attention_2
-  max_length: 12288
-
-sampler:
-  max_model_len: 12288
-  gpu_memory_utilization: 0.8
-  max_num_seqs: 32
-  max_num_batched_tokens: 16384
-  enforce_eager: true
-
-tq:
-  polling_mode: true
-  storage_units: 2
-
-scheduler:
-  # Keep scheduler settings identical to the four-context experiment.
-  rollout: {policy: round_robin, max_consecutive_units: 1}
-  advantage: {policy: oldest_partition, max_consecutive_units: 1}
-  train: {policy: round_robin, max_consecutive_units: 1}
-
-evaluation:
-  enabled: true
-  interval: 5
-  batch_size: 16
-  sampling_params:
-    max_tokens: 8192
-    temperature: 0.0
-    top_p: 1.0
-
-lora:
-  target_modules: all-linear
-  r: 16
-  alpha: 32
-  dropout: 0.05
-  learning_rate: 1.0e-6
-
-loss:
-  cls: GRPOLoss
-  epsilon: 0.2
-
-lora_contexts:
-  - tenant_id: tenant_single_dapo
-    training_run_id: dapo_single_tenant
-    adapter_name: dapo_single_lr1e6_g4_lora
-    reward:
-      class_path: twinkle.reward.DAPOMathReward
-      kwargs:
-        max_response_length: ${...rollout.max_tokens}
-        # Intentionally identical to tenant_lr1e6_g4 in the four-context sweep.
-        overlong_buffer_length: 4096
-        overlong_penalty_factor: 1.0
-        score_tail_chars: 300
-    dataset:
-      dataset_id: ${oc.env:DAPO_SINGLE_DATASET_ID,ms://BytedTsinghua-SIA/DAPO-Math-17k}
-      subset_name: default
-      split: train
-      data_num: 500
-      max_length: 4096
-      processor: DAPOMathProcessor
-    eval_dataset:
-      name: aime2024
-      dataset_id: ${oc.env:AIME2024_DATASET_ID,ms://Maxwell-Jia/AIME_2024}
-      subset_name: default
-      split: train
-      data_num: null
-      max_length: 4096
-      processor: AIME2024Processor
-      reward:
-        class_path: twinkle.reward.BoxedMathAccuracyReward
-    rollout:
-      batch_size: 16
-      num_generations: 8
-      max_tokens: 8192
-      temperature: 1.0
-      top_p: 0.95
-    train:
-      learning_rate: 1.0e-6
-      mini_batch_size: 32
-      micro_batch_size: 1
-      dynamic_batching: false
-
-
-# CUDA_VISIBLE_DEVICES=0,1,2 \
-# MODEL_ID=/nas/disk1/Qwen3-4B \
-# DAPO_SINGLE_DATASET_ID=/path/to/the_same_dapo_math_500.parquet \
-# TWINKLE_SEED=42 \
-# python3 cookbook/rl/async_multi_lora_grpo.py \
-#   --config cookbook/rl/async_single_lora_dapo_grpo.yaml
