--- cookbook/rl/async_single_lora_gsm8k_areal.yaml
@@ -1,190 +1,0 @@
-runtime:
-  run_id: async_single_lora_gsm8k_areal
-  model_id: ${oc.env:MODEL_ID,/nas/disk1/Qwen3-4B}
-  mode: ray
-  model_gpus: 1
-  sampler_gpus: 1
-  sampler_tp: 1
-  sampler_max_loras: 1
-  seed: 1
-  max_staleness: 0
-  max_steps: null
-  allow_partial_rollout: false
-  rollout_max_retries: 2
-  rollout_retry_delay_s: 0.5
-  keep_adapter_versions: 2
-  output_dir: output/async_single_lora_gsm8k_areal
-
-metrics:
-  enabled: true
-  drain_interval_s: 1.0
-  queue_capacity: 10000
-  close_timeout_s: 10
-  jsonl:
-    enabled: true
-    path: outputs/async_rl/single_lora_gsm8k_areal_metrics.jsonl
-    summary_path: outputs/async_rl/single_lora_gsm8k_areal_summary.json
-    batch_size: 64
-    flush_interval_s: 2.0
-  swanlab:
-    enabled: false
-    mode: local
-    project: twinkle-rl
-    name: ${runtime.run_id}
-    log_dir: outputs/swanlab
-
-template:
-  cls: Template
-  enable_thinking: true
-
-model:
-  strategy: native_fsdp
-  attn_implementation: flash_attention_2
-  fsdp_config:
-    reshard_after_forward: true
-  sequence_parallel_size: 1
-  padding_free: false
-  max_length: 2048
-
-sampler:
-  max_model_len: 2048
-  gpu_memory_utilization: 0.8
-  max_num_seqs: 64
-  enforce_eager: false
-
-rollout_output:
-  enabled: true
-  output_dir: ${runtime.output_dir}/rollouts
-  include_token_ids: false
-
-tq:
-  polling_mode: true
-  storage_units: 2
-
-scheduler:
-  rollout: {policy: round_robin, max_consecutive_units: 1}
-  advantage: {policy: oldest_partition, max_consecutive_units: 1}
-  train: {policy: sticky, max_consecutive_units: null}
-
-evaluation:
-  enabled: false
-  interval: 10
-  batch_size: 16
-  sampling_params:
-    max_tokens: 2048
-    temperature: 0.6
-    top_p: 1.0
-
-lora:
-  target_modules: all-linear
-  r: 16
-  alpha: 16
-  dropout: 0.0
-  learning_rate: 1.7e-5
-
-loss:
-  cls: GRPOLoss
-  epsilon: 0.2
-
-lora_contexts:
-  - tenant_id: tenant_single
-    training_run_id: gsm8k_areal_pk
-    adapter_name: gsm8k_areal_pk_lora
-    reward:
-      class_path: twinkle.reward.MathVerifyAccuracyReward
-    dataset:
-      dataset_id: ${oc.env:GSM8K_TRAIN_DATASET_ID}
-      subset_name: main
-      split: train
-      data_num: 2000
-      max_length: 1024
-      processor: AReaLGSM8KProcessor
-    eval_dataset:
-      name: gsm8k/test
-      dataset_id: ${oc.env:GSM8K_TEST_DATASET_ID}
-      subset_name: main
-      split: test
-      data_num: null
-      max_length: 1024
-      processor: AReaLGSM8KProcessor
-      reward:
-        class_path: twinkle.reward.MathVerifyAccuracyReward
-    rollout:
-      batch_size: 16
-      num_generations: 4
-      max_tokens: 1024
-      temperature: 1.0
-      top_p: 1.0
-    train:
-      mini_batch_size: 64
-      micro_batch_size: 64
-      dynamic_batching: true
-      max_tokens_per_micro_batch: 4096
-      packing_algorithm: ffd
-
-
-# export VIRTUAL_ENV=/opt/.venv
-# export PATH=/opt/.venv/bin:$PATH
-# export CUDA_VISIBLE_DEVICES=2,3
-# export AREAL_MODEL_PATH=/nas/disk1/Qwen3-4B
-# export AREAL_ADMIN_API_KEY="$(
-#   python3 -c 'import secrets; print(secrets.token_urlsafe(32))'
-# )"
-# export AREAL_TRIAL="pk-$(date +%Y%m%d-%H%M%S)"
-
-# python3 examples/math/gsm8k_rl.py \
-#   --config examples/math/gsm8k_grpo_lora.yaml \
-#   scheduler.type=local \
-#   +rollout.agent.admin_api_key="$AREAL_ADMIN_API_KEY" \
-#   seed=1 \
-#   cluster.n_nodes=1 \
-#   cluster.n_gpus_per_node=2 \
-#   actor.path="$AREAL_MODEL_PATH" \
-#   +actor.attn_impl=flash_attention_2 \
-#   rollout.backend=vllm:d1 \
-#   actor.backend=fsdp:d1 \
-#   train_dataset.path=/model/ljl/project/data/gsm8k \
-#   valid_dataset.path=/model/ljl/project/data/gsm8k \
-#   train_dataset.batch_size=16 \
-#   train_dataset.shuffle=false \
-#   total_train_epochs=1 \
-#   +total_train_steps=125 \
-#   rollout.consumer_batch_size=16 \
-#   rollout.max_concurrent_rollouts=16 \
-#   rollout.max_head_offpolicyness=0 \
-#   gconfig.n_samples=4 \
-#   gconfig.max_new_tokens=1024 \
-#   gconfig.max_tokens=2048 \
-#   gconfig.temperature=1.0 \
-#   +gconfig.top_p=1.0 \
-#   +actor.mb_spec.n_mbs=64 \
-#   ++actor.mb_spec.max_tokens_per_mb=null \
-#   actor.use_lora=true \
-#   actor.lora_rank=16 \
-#   actor.lora_alpha=16 \
-#   actor.optimizer.lr=1.7e-4 \
-#   actor.optimizer.weight_decay=0.01 \
-#   actor.optimizer.lr_scheduler_type=constant \
-#   actor.eps_clip=0.2 \
-#   actor.ppo_n_minibatches=1 \
-#   actor.reward_scaling=1.0 \
-#   actor.reward_bias=0.0 \
-#   actor.adv_norm=null \
-#   actor.kl_ctl=0.0 \
-#   actor.recompute_logprob=false \
-#   actor.use_decoupled_loss=false \
-#   actor.rejection_sampling=null \
-#   ++vllm.max_model_len=2048 \
-#   ++vllm.gpu_memory_utilization=0.8 \
-#   ++vllm.max_num_seqs=64 \
-#   ++vllm.max_loras=1 \
-#   ++vllm.enforce_eager=false \
-#   evaluator.freq_epochs=null \
-#   evaluator.freq_steps=null \
-#   evaluator.freq_secs=null \
-#   trial_name="$AREAL_TRIAL" \
-#   +stats_logger.tensorboard.path="/tmp/areal/tensorboard/gsm8k-grpo/$AREAL_TRIAL" \
-#   cluster.fileroot=/nas/disk1/areal-experiments \
-#   saver.freq_epochs=null \
-#   saver.freq_steps=125 \
-#   saver.freq_secs=null
